Create LandsatSSTcalculator.js
This code helps to create a seasonal SST using Landsat imagery. It selects all the available data over a given area and time range to get an averaged SST image, using Landsat 5,7, and 8. The outputs are a CSV file containing the names and properties of the images used in the average, and an image that can be exported as an asset for future use in GEE.